requirements.txt
# API
fastapi==0.135.1
uvicorn==0.41.0
python-multipart==0.0.22

# ML
torch==2.10.0
torchvision==0.25.0
transformers==5.3.0
safetensors==0.7.0
tokenizers==0.22.2
accelerate>=0.20.0
huggingface-hub>=0.16.0

# Utils
pillow==12.1.1
numpy==2.4.3
requests==2.32.5
tqdm==4.67.3
filelock==3.25.2